refactor: remove auto-loop.ts barrel, import from ./auto/* directly
The auto-loop.ts file was a pure re-export barrel re-exporting from
./auto/loop.js, ./auto/resolve.js, ./auto/run-unit.js, ./auto/detect-stuck.js,
./auto/infra-errors.js. Removing this indirection layer.

Updates 6 importers to use direct paths:
- src/resources/extensions/gsd/auto.ts
- src/resources/extensions/gsd/bootstrap/agent-end-recovery.ts
- tests/auto-loop.test.ts (also drops the export-validation test that
  asserted the barrel itself existed)
- tests/custom-engine-loop-integration.test.ts
- tests/journal-integration.test.ts (6 dynamic imports)
- tests/worktree-path-injection.test.ts (1 dynamic import using .ts ext)

No behavior change. Pure import-path refactor.
